Fix inorrect copying of skb head
We should not copy value of skb->tail_lock for cloned skb,
it should be zero for a new skb. For example in `skb_segment`
function we clone skb and then use skb_put on the clonned skb,
that leads to BUG_ON if skb->tail_lock is not zero.