Reliable memory reserve
This patch consist of two parts, first is adjusting
of KASLR for reserve huge amount of continuous pages.
We clamp the max offset of KASLR's point where kernel
is unpacked. For instance, there is the memory region
14G we need to reserve 12G, max offset where kernel may
be placed is 14G-12G = 2G.

The second part is memory reserve using memblock API
and then preparing of the huge pages. This is more
reliable than using buddy for such allocations.
However, on NUMA setup sometimes we still have
issues with KASLR, when trying to allocate 70-80% of
the memory. With disabled KASLR allocation more
reliable on NUMA machine.