Fix warning in `skb_release_head_state`.
This commit is a copy of `f8321fa75102246d7415a6af441872f6637c93ab`
commit from linux kernel source code.
The issue arises because virtio is assuming it's running in NAPI
context even when it's not, such as in the netpoll case.
(see trace from #2230 in Tempesta or cases from appropriate
linux kernel commit).

To resolve this, modify virtnet_poll_tx() to only set NAPI when budget
is available. Same for virtnet_poll_cleantx(), which always assumed that
it was in a NAPI context.